Rename package name to bgames
