+++ src/main/java/com/arexe/bgames/BoardGamesApplication.java
@@ -1,0 +1,13 @@
+package com.arexe.bgames;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BoardGamesApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BoardGamesApplication.class, args);
+	}
+
+}
