+++ src/main/java/com/arexe/bgames/config/WebMvcConfig.java
@@ -1,0 +1,15 @@
+package com.arexe.bgames.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    @Bean
+    public BCryptPasswordEncoder pwdEncrypt(){
+        return new BCryptPasswordEncoder();
+    }
+}
