+++ src/main/java/com/arexe/bgames/controllers/LoginController.java
@@ -1,0 +1,13 @@
+package com.arexe.bgames.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginController {
+
+    @GetMapping(value = "/login")
+    public String loginPage(){
+        return "login";
+    }
+}
