+++ src/main/java/com/arexe/bgames/controllers/MainPageController.java
@@ -1,0 +1,13 @@
+package com.arexe.bgames.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainPageController {
+
+    @GetMapping(value = {"/", "/index"})
+    public String showMainPage() {
+        return "index";
+    }
+}
