+++ src/main/java/com/arexe/bgames/repository/RoleRepository.java
@@ -1,0 +1,12 @@
+package com.arexe.bgames.repository;
+
+import com.arexe.bgames.entity.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Integer> {
+
+    Role findByRole(String role);
+
+}
