+++ src/main/java/com/arexe/bgames/service/UserService.java
@@ -1,0 +1,12 @@
+package com.arexe.bgames.service;
+
+import com.arexe.bgames.entity.User;
+
+public interface UserService {
+
+    void saveUser(User user);
+    User findUserByEmail(String email);
+    User findUserByLogin(String login);
+    void updateUserPassword(String newPassword, String email);
+    void updateUserProfile(String login, String firstName, String lastName, String email);
+}
