+++ src/main/java/com/arexe/bgames/utils/UserUtils.java
@@ -1,0 +1,17 @@
+package com.arexe.bgames.utils;
+
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+
+public class UserUtils {
+
+    public static String getLoggedUser() {
+        String username = null;
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if(!(auth instanceof AnonymousAuthenticationToken)) {
+            username = auth.getName();
+        }
+        return username;
+    }
+}
