+++ src/main/java/com/arexe/bgames/validators/PatternChecker.java
@@ -1,0 +1,13 @@
+package com.arexe.bgames.validators;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class PatternChecker {
+
+    public static boolean check(String pattern, String input) {
+        Pattern p = Pattern.compile(pattern);
+        Matcher m = p.matcher(input);
+        return m.matches();
+    }
+}
