--- src/main/java/com/arexe/portal/config/WebMvcConfig.java
@@ -1,15 +1,0 @@
-package com.arexe.portal.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    @Bean
-    public BCryptPasswordEncoder pwdEncrypt(){
-        return new BCryptPasswordEncoder();
-    }
-}
