--- src/main/java/com/arexe/portal/controllers/LoginController.java
@@ -1,13 +1,0 @@
-package com.arexe.portal.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-
-    @GetMapping(value = "/login")
-    public String loginPage(){
-        return "login";
-    }
-}
