--- src/main/java/com/arexe/portal/repository/RoleRepository.java
@@ -1,12 +1,0 @@
-package com.arexe.portal.repository;
-
-import com.arexe.portal.entity.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-
-    Role findByRole(String role);
-
-}
