--- src/main/java/com/arexe/portal/service/AdminService.java
@@ -1,14 +1,0 @@
-package com.arexe.portal.service;
-
-import com.arexe.portal.entity.User;
-
-import java.util.List;
-
-public interface AdminService {
-
-    List<User> getUserList();
-    List<User> findUsersByName(String name);
-    User getUserById(int id);
-    void updateUserStatus(int id, int roleNumber, int active);
-    void deleteUserById(int id);
-}
