+++ src/main/resources/database/portal.sql
@@ -19,7 +19,7 @@ SET time_zone = "+00:00";
 /*!40101 SET NAMES utf8mb4 */;
 
 --
--- Baza danych: `portal`
+-- Baza danych: `bgames`
 --
 
 -- --------------------------------------------------------
