+++ src/test/java/com/arexe/bgames/BoardGamesApplicationTests.java
@@ -1,0 +1,13 @@
+package com.arexe.bgames;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class BoardGamesApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
