--- src/test/java/com/arexe/portal/PortalApplicationTests.java
@@ -1,13 +1,0 @@
-package com.arexe.portal;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class PortalApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
